mixins improvements #103
