+++ lib/mixins/browser_queries.js
@@ -1,7 +1,7 @@
 "use strict";
 
-module.exports = function BrowserBaseMixin(s) {
-    return class BrowserBase extends s {
+module.exports = function BrowserQueriesMixin(s) {
+    return class BrowserQueries extends s {
 
         evaluate(cb, ...args) {
             this._failIfNotLoaded();
@@ -33,6 +33,32 @@ module.exports = function BrowserBaseMixin(s) {
         queryXPath(xPath) {
             this._failIfNotLoaded();
             return this.page.$x(xPath);
+        }
+
+        findByText(text, optionalText) {
+            this._failIfNotLoaded();
+            const xPathText = optionalText || text;
+            const xPath = `//*[text()='${xPathText}']`;
+            if(optionalText) {
+                return this.query(text).then((elem) => {
+                    return this._subQueryXpath(elem, xPath);
+                });
+            } else {
+                return this.queryXPath(xPath);
+            }
+        }
+
+        findByTextContaining(text, optionalText) {
+            this._failIfNotLoaded();
+            const xPathText = optionalText || text;
+            const xPath = `//*[contains(text(),'${xPathText}')]`;
+            if(optionalText) {
+                return this.query(text).then((elem) => {
+                    return this._subQueryXpath(elem, xPath);
+                });
+            } else {
+                return this.queryXPath(xPath);
+            }
         }
 
         _subQueryXpath(parent, selector) {
